Use new suppressable subject and move to new folder
