+++ AudiosAmigo.Droid/Observables/ObservableOnItemLongClickListener.cs
@@ -3,7 +3,7 @@ using System.Reactive.Subjects;
 using Android.Views;
 using Android.Widget;
 
-namespace AudiosAmigo.Droid
+namespace AudiosAmigo.Droid.Observables
 {
     class ObservableOnItemLongClickListener<T> :
         Java.Lang.Object, AdapterView.IOnItemLongClickListener, IObservable<T>
