+++ AudiosAmigo.Droid/Observables/ObservableSeekBarListener.cs
@@ -1,36 +1,22 @@
 using System;
 using System.Reactive.Linq;
 using System.Reactive.Subjects;
-using System.Threading;
-using System.Threading.Tasks;
 using Android.Widget;
 
-namespace AudiosAmigo.Droid
+namespace AudiosAmigo.Droid.Observables
 {
     public class ObservableSeekBarListener : Java.Lang.Object, SeekBar.IOnSeekBarChangeListener, IObservable<int>
     {
         private readonly Subject<int> _progressUpdate = new Subject<int>();
 
         private int _progress;
-
-        private int _suppressCounter;
         
         public ObservableSeekBarListener(SeekBar seekbar)
         {
             _progress = seekbar.Progress;
             seekbar.SetOnSeekBarChangeListener(this);
         }
-
-        public void Suppress(int delay)
-        {
-            Interlocked.Increment(ref _suppressCounter);
-            Task.Run(async () =>
-            {
-                await Task.Delay(delay);
-                Interlocked.Decrement(ref _suppressCounter);
-            });
-        }
-
+        
         public void OnProgressChanged(SeekBar seekBar, int progress, bool fromUser)
         {
             Emit(seekBar);
@@ -48,12 +34,9 @@ namespace AudiosAmigo.Droid
 
         private void Emit(SeekBar seekBar)
         {
-            if (_suppressCounter <= 0)
+            if (_progress != seekBar.Progress)
             {
-                if (_progress != seekBar.Progress)
-                {
-                    _progressUpdate.OnNext(seekBar.Progress);
-                }
+                _progressUpdate.OnNext(seekBar.Progress);
             }
             _progress = seekBar.Progress;
         }
