+++ AudiosAmigo.Windows/Observables/ObservableAudioSessionEvents.cs
@@ -1,22 +1,19 @@
 using System;
-using System.Reactive.Subjects;
-using System.Threading;
 using System.Reactive.Linq;
-using System.Threading.Tasks;
-using NAudio.CoreAudioApi.Interfaces;
+using System.Reactive.Subjects;
 using NAudio.CoreAudioApi;
+using NAudio.CoreAudioApi.Interfaces;
 
-namespace AudiosAmigo.Windows
+namespace AudiosAmigo.Windows.Observables
 {
     public class ObservableAudioSessionEvents : IAudioSessionEventsHandler, IObservable<Tuple<float, bool>>
     {
-        private readonly Subject<Tuple<float, bool>> _subject = new Subject<Tuple<float, bool>>();
+        private readonly SupressSubject<Tuple<float, bool>> _subject = 
+            new SupressSubject<Tuple<float, bool>>(new Subject<Tuple<float, bool>>());
 
         private float _volume;
 
         private bool _mute;
-
-        private int _suppressCounter;
 
         public ObservableAudioSessionEvents(AudioSessionControl session)
         {
@@ -37,12 +34,7 @@ namespace AudiosAmigo.Windows
         
         public void Suppress(int delay)
         {
-            Interlocked.Increment(ref _suppressCounter);
-            Task.Run(async () =>
-            {
-                await Task.Delay(delay);
-                Interlocked.Decrement(ref _suppressCounter);
-            });
+            _subject.Suppress(delay);
         }
 
         public void OnChannelVolumeChanged(uint channelCount, IntPtr newVolumes, uint channelIndex){}
@@ -65,12 +57,9 @@ namespace AudiosAmigo.Windows
 
         public void OnVolumeChanged(float volume, bool mute)
         {
-            if (_suppressCounter <= 0)
+            if (Math.Abs(_volume - volume) > 0.001 || _mute != mute)
             {
-                if (Math.Abs(_volume - volume) > 0.001 || _mute != mute)
-                {
-                    _subject.OnNext(Tuple.Create(volume, mute));
-                }
+                _subject.OnNext(Tuple.Create(volume, mute));
             }
             _volume = volume;
             _mute = mute;
