+++ AudiosAmigo.Windows/Observables/ObservableEndpointVolumeNotification.cs
@@ -1,21 +1,18 @@
 using System;
 using System.Reactive.Linq;
 using System.Reactive.Subjects;
-using System.Threading;
-using System.Threading.Tasks;
 using NAudio.CoreAudioApi;
 
-namespace AudiosAmigo.Windows
+namespace AudiosAmigo.Windows.Observables
 {
     public class ObservableEndpointVolumeNotification : IObservable<Tuple<float, bool>>
     {
-        private readonly Subject<Tuple<float, bool>> _subject = new Subject<Tuple<float, bool>>();
+        private readonly SupressSubject<Tuple<float, bool>> _subject =
+            new SupressSubject<Tuple<float, bool>>(new Subject<Tuple<float, bool>>());
 
         private float _volume;
 
         private bool _mute;
-
-        private int _suppressCounter;
 
         public ObservableEndpointVolumeNotification(AudioEndpointVolume endpointVolume)
         {
@@ -39,22 +36,14 @@ namespace AudiosAmigo.Windows
 
         public void Suppress(int delay)
         {
-            Interlocked.Increment(ref _suppressCounter);
-            Task.Run(async () =>
-            {
-                await Task.Delay(delay);
-                Interlocked.Decrement(ref _suppressCounter);
-            });
+            _subject.Suppress(delay);
         }
 
         public void OnVolumeChanged(float volume, bool mute)
         {
-            if (_suppressCounter <= 0)
+            if (Math.Abs(_volume - volume) > 0.001 || _mute != mute)
             {
-                if (Math.Abs(_volume - volume) > 0.001 || _mute != mute)
-                {
-                    _subject.OnNext(Tuple.Create(volume, mute));
-                }
+                _subject.OnNext(Tuple.Create(volume, mute));
             }
             _volume = volume;
             _mute = mute;
